+++ CHANGELOG.md
@@ -36,6 +36,7 @@
 
 - wrong return-code in one error-case of the create-train-task endpoint was fixed
 - fixed offset in the output of the timeseries training from datasets
+- download of dataset with a too big number of requested rows and offset was fixed to handle this cases correctly
 
 ### Removed